Add files via upload
I have modified the KeyMatrix and LCD header files to have #defines in the header instead of the source. Also sorted out the problem I was having with some parts of the LCD. On top of this I sorted the Thermometer into a driver and have included a main interacting with these three parts.